Crate test_getsockname

Structs

GetsocknameArguments

Functions

check_getsockname_call

get_tests

inet_sockaddr_check_equal

main

socket_helper

test_after_close

Test getsockname after closing the bound socket.

test_autobound_socket

Test getsockname using an autobound socket.

test_bound_connected_socket

Test getsockname using a bound and connected socket.

test_bound_socket

Test getsockname using a bound socket.

test_implicit_bind

Test getsockname using a listening socket without binding (an implicit bind).

test_invalid_fd

Test getsockname using an argument that cannot be a fd.

test_non_existent_fd

Test getsockname using an argument that could be a fd, but is not.

test_non_socket_fd

Test getsockname using a valid fd that is not a socket.

test_null_addr

Test getsockname using a valid fd, but with a NULL address.

test_null_len

Test getsockname using a valid fd and address, a NULL address length.

test_short_len_inet

Test getsockname using a valid fd and address, but an address length that is too small.

test_unbound_socket

Test getsockname using an unbound socket.

test_zero_len

Test getsockname using a valid fd and address, but an address length of 0.